just two warnings casts for linux Submitted by: Aram Mirzadeh (awm@qosina.com)
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@76327 13f79535-47bb-0310-9956-ffa450edef68

+++ src/mod_proxy.c
@@ -1813,7 +1813,7 @@ cache_update(struct cache_req *c, array_header *resp_hdrs,
     if (c->fp != NULL)
     {
 	pclosef(r->pool, c->fp->fd);
-	c->fp->fd = NULL;
+	c->fp->fd =(int) NULL;
     }
     c->version = 0;
     sec2hex(0, buff+27);
